Initial commit
This is the help from the latest Dev branch of the HyperHelp code
stream, which contains all of the help files and the table of contents
for the Sublime API help.

This will eventually become its own package in its own repository (not
this one), but this will allow us to track changes.