Do some cleanup and setup maintenance
Now that the restriction as to where the hyperhelp.json index file
goes have been relaxed, put it into the doc directory along with the
help files themselves.

Also include a dependencies file that tells PackageControl that we
require the hyperhelp dependency in order to operate, and set up a
command file that makes opening the SnAPI help more direct.