Update to use the new help system
The index and help files have been updated to work with the new
hyperhelp prototype developed during Devember 2017.

This is just formatting changes to use the new index format. There is
still a lot to do; for example the captions on topics are not set up
yet and not all links have been verified, plus the TOC can be
simplified greatly using the new API available.