Merge Devember 2017 rework into this repository
During December 2017 I rewrote my hyperhelp project from the ground
up.

As a part of that project I also worked on the contents of this
package as well, as the help index needed updating to support the
newer HyperHelp version. The help was also updated to take better
advantage of the new hyperhelp features.

This commit represents the code as of the end of the last day of
Devember, squashed into a single commit.