Update API help up to Sublime 3.1.1
Since the creation of the initial SnAPI help documentation, some
enhancements have been made to the API and the documentation.

In particular the new TextInputHandler and ListInputHandler classes
were added to the API to allow for new input features in the command
palette, and the ViewEventListener was enhanced as well.

As a part of this change, some of the typos that existed in the
original help files (as taken from the web page) were also fixed up.