Get prepared for multi-package help
Changes are being made in the core of HyperHelp to allow a single
package to contain more than one help index, allowing for the help of
multiple packages to be contained in a single package for
distribution.

To prepare for that, and wanting to have other Sublime help in this
package, we're changing the location of the current help files to
track what help package they're for.

Along with this change the document root is also modified, and the new
"package" top level key is set to specify the name of the package the
help is for.

As yet this is not a feature of hyperhelp, so this makes the index
essentially broken unless you also have the newer version of the code,
although these commits will not be made public until all packages have
been updated accordingly.