Include the first parts of the build system docs
This includes as proof of concept the first parts of the build system
documentation. This is somewhere less than half done as far as
transcribing the help information is concerned.

At this point the index is still not complete pending all of the help
files and topics being in place. Once all of the information is
available we may want to shuffle the files around to make them more
navigable since at the moment the usage page has a lot of table
content compared to the text.

Also needed will be a table of contents.