Enable jumping to build help from build files
This enhances the command for opening the SnAPI help so that if it is
invoked from inside of a build file (even an unsaved template), the
help index for the build help is opened.

In future this could be enhanced by trying to find the closest key in
the help and jumping directly to that topic, optionally going to the
index if the key could not be found, or something along those lines.