Finalize the initial build system help content
This finishes up creating the files that contain the help for build
systems as derived from the official documentation.

There is undoubtedly more work to clean this up, and in particular the
index is not complete yet because it doesn't list an appropriate table
of contents, but we're getting there.

Note that at the current time this requires the experimental hyperhelp
syntax because it uses markdown style code fences to specify
languages.