Enhance the SnAPI lookup command
This enhances the SnAPI lookup command so that it works the same way
in build files as it does in plugins; the word under the cursor is
used as the initial topic, and if it is a topic in the index, that
topic is opened directly.

This also includes the notion of a default package to display the
index for in cases where the current file is not a plugin or a build
file, which is currently hard coded to be the root topic in the
SublimeAPI help package.

This should probably at some level be configurable by the user.