Expand the aliases available in the API docs
The API endpoints are all added with topics that contain their class
short names, such as `view.name`. This allows the topic to carry
information as to context, which is important because API endpoints
are not unique (all `Command` subclasses contain a `run` method, for
example).

This change adds aliases to all API endpoint topics that are not
ambiguous so that looking up just `name` will find `view.name`.

This enhances the utility of the SnAPI lookup function, which extracts
the word under the cursor and thus has a potentially better chance of
finding an appropriate topic.