Enhance how the appropriate help package is picked
This modifies how SnAPI determines what help file it should open for its
context command by first checking the scope of the file and then falling
back to the extension on the file if the scope is ambiguous/unknown.

This is superior to the prior mechanism when PackageDev is installed
because PD has syntaxes with more unique scopes in them whereas the core
uses `source.json`, requiring a file name in order to operate.