Clean up some argument names in the API
Per the list in https://github.com/SublimeTextIssues/Core/issues/2290
this makes some changes to the parameter names of some API functions to
better match what the real underlying parameter in the methods are
called.

The main impetus for this change is to change some of the optional
arguments to API functions that you sometimes use keyword arguments for
which are misnamed in the documentation.

Along with that, some parameter names have also been renamed where my
(probably faultly) judgment deems the replacement with the underlying
actual name makes the document a bit more readable.

In a similar vein, not all such items in the issue above have been
modified because in many cases the actual name makes the documentation
less readable (again in my probably faulty judgement), not more readable
as one would hope.

An example of such is the `fold` and `unfold` API functions, which take
actual arguments named `x` but which the document calls `region` or
`[regions]`, which is much more readable for the purposes of knowing
what it does.