Include undocumented API endpoints in sublime.txt
Using the same issue as previously referenced, this adds in undocumented
API endpoints in the Sublime module. Where possible the endpoints are
documented as to their function.

Each such endpoint is flagged as being undocumented as well as (in some
cases) deprecated, so that the user knows what they may or may not be
getting themselves into.