Update help to support new features
This updates all help files to support new features that have been added
in the core help system.

In particular this means things like having anchors with no topic to aid
in navigation in-file, changing some header hidden anchors to include
the full text of the heading, and so on.