Include new mention to dev build docs
The latest version of the color scheme documentation links to a page
that outlines the color scheme support in dev builds, which is different
(in theory) from stable builds.

Here we are just including that reference as well, with a link to the
relevant page. It looks at first blush like this is a complete extra set
of documentation instead of just patch notes, and at least currently it
doesn't seem like a good idea to try and include that as well since it
will be mostly the same anyway.