Update and clean up help files
While updating the contents of the color scheme help file to include the
recently documented `invisibles` key, I realized that there were still a
lot of links and anchors using the old syntax even though the linter has
been updated to treat them as errors.

Under the assumption that I did that because we wanted to be sure to do
a clean up that perhaps never happened due to Devember spinning up, that
is now done so that things are clean going forward.