Mark deleted plugins as such until official update
The general.txt page contains a list of default plugins that you can use
to see how Sublime plugins work, but in build 3189 three of them were
removed and replaced with equivalent commands in the core.

As such, here they have been marked to act as a reminder to anyone that
is using `hyperhelp` on dev builds, since those files will refuse to
open due to their no longer existing.