Inline the development documentation changes
This change syncs up the current documentation by also referencing the
points in the color scheme documentation where the development versions
alter the documented behaviour of stable builds.

The implication here is that such items are clearly flagged as
`[DEV BUILDS ONLY]` to indicate this, and that this represents the most
recent development build series.

With this change, it behooves us to ensure that we don't lag behind
official documentation changes too far.