Include a method from the new dev series
This includes a preliminary documentation entry for the new API endpoint
for setting the minidiff target content in the view.

This is here pending official documentation changing to include this, at
which point we will be able to remove the annotations and fix the help
text to be more "official".