Include the new API additions in build 3190
This build introduces a new API endpoint that resets the mini diff
reference document back to the current contents of the buffer, which
turns off all of the diff content.

This is just a shortcut to setting the reference document to the current
view contents, but is clearer in its intention.