Update the documentation for version 3.2
On this historic day, Sublime Text 3.2 was officially released. As such
this change goes through existing documentation to remove the
annotations that marked items as being for dev builds only, since they
are now currently stable.

There will likely be an upcoming change for this as it was missed by the
official team that some of the plugins that are referenced in the API
documentation as example plugins were removed in build 3189.

When that change goes live (which will probably be in the next couple of
days, if I had to guess), another update will be pushed for this.