Remove redacted plugins from the plugin list
This change removes from the list of example plugins the three that were
removed in build 3189 (because their functionality moved to the core for
speed).

This was a documentation oversight during the update of the dev docs for
build 3200 that I mentioned to Will on the discord and which has now
been resolved.