Update all help files to use content headers
This uses the new syntax that allows us to use markdown style header
comments in order to get the symbol list in a help file to show the
top level headings, which allows for easier traversal of the page.

This is particularly noticeable in the help files for things like the
color schemes, which are loaded with anchors.