Return the few instances of `iff` that were elided
There are five places where the official help uses `iff` which I
originally changed to `if` during the transcription. However this was
actually intentional and means "if and only if".

In the context of how this is used, I'm not sure that this is any more
clear or saying something different than just using `iff`, but in the
spirit of not modifying the official docs where possible, this change
replaces those items.