Enhance the topic selection for cursor location
This implements a new method for grabbing the word that is currently
under the caret. The `expand_by_class()` API method on `view` sometimes
moves past where one might expect it to.

Here we directly reject grabbing a word unless the caret is on one, and
we carefully extract just the bounds and skip any word separators that
might be set.