Update copyright messages
After creating the videos (because all errors are noticed after a long
series of recording and editing sessions; the universe hates me) I
realized that the help for the color scheme and build systems said in
the copyright in the page footer that the help was API help.

This is changed to a more specific "Sublime Text Help" instead, which is
clearer generally and better matches the source pages.