Fix some slight documentation changes
Include a couple of documentation changes for things which were somehow
missed on update, which includes the ListInputHandler's `description()`
method call signature and the fact that the `cross` icon is no longer
valid for `add_regions()`.

I'm not sure how these slipped past the automatic detection; I'm
guessing they happened at the same time as a more extensive change, such
as to  the CSS or fonts used by the official docs.