Update the build documentation
This is a refresh of the build documentation from the new documentation
set. This includes some small wording changes and now (since it's
possible where it wasn't previously) we can link out to some of the API
documentation as well when referring to custom targets.

There is a link to file patterns which is not currently available in
SnAPI, but which should be. For now it links to the external page.