Include new dialog and color scheme options
This includes the new dialog methods that allow plugins to prompt the
user to select files to open and save so that they can capture paths,
as well as the ability to select folders.

In addition, some new font options were added to font styles in the form
of different forms of underlined text.

We're also piggybacking in a small change to the view help because
`indexed_symbols` was incorrectly labeled as `indexed_symbol`.