Fix a deprecation notice.
The `view.set_syntax_file()` API endpoint is deprecated in favor of a
new `view.assign_syntax()` method (and is currently implemented by
calling the new method).

This is not a new thing by any stretch and has been in place since
the first SnAPI versions, but the help for this element was incorrectly
linking to itself as the thing to use, which is not correct.